# Add QHRF Harmonic Resonance Pass - UnitaryHACK 2025 Bounty #1