--- a/ucc/transpilers/qhrf_pass.py
+++ b/ucc/transpilers/qhrf_pass.py
@@ -6,7 +6,7 @@ unprecedented quantum circuit optimization through resonance-based
 gate elimination and coherence stabilization.
 
 Experimental validation on IBM Brisbane achieved:
-- 100% CNOT gate reduction  
+- 100% CNOT gate reduction
 - 99.6% quantum coherence retention across 40 qubits
 - 85% circuit depth reduction after transpilation
 
@@ -27,30 +27,30 @@ import numpy as np
 class QHRFHarmonicResonancePass(TransformationPass):
     """
     Quantum Harmonic Resonance Framework (QHRF) Compiler Pass
-    
+
     This pass implements a novel quantum optimization technique based on
-    harmonic resonance theory. It replaces traditional force-mediated 
+    harmonic resonance theory. It replaces traditional force-mediated
     quantum interactions (CNOT gates) with resonance-stabilized patterns
     that maintain logical equivalence while dramatically improving
     circuit efficiency.
-    
+
     Key Features:
     - Complete CNOT gate elimination through resonance substitution
-    - Enhanced quantum coherence via harmonic stabilization  
+    - Enhanced quantum coherence via harmonic stabilization
     - Superior transpilation efficiency on quantum hardware
     - Experimentally validated on IBM quantum processors
-    
+
     Performance Improvements:
     - 100% reduction in 2-qubit gate counts
     - Up to 85% circuit depth reduction after transpilation
     - Enhanced quantum state coherence retention
     - Faster compilation times
     """
-    
+
     def __init__(self, apply_coherence_stabilization: bool = True):
         """
         Initialize the QHRF Harmonic Resonance Pass
-        
+
         Args:
             apply_coherence_stabilization: Whether to apply additional
                 coherence stabilization operations
@@ -58,86 +58,89 @@ class QHRFHarmonicResonancePass(TransformationPass):
         super().__init__()
         self.apply_coherence_stabilization = apply_coherence_stabilization
         self._optimization_stats = {
-            'cnot_gates_eliminated': 0,
-            'resonance_patterns_applied': 0,
-            'coherence_operations_added': 0
+            "cnot_gates_eliminated": 0,
+            "resonance_patterns_applied": 0,
+            "coherence_operations_added": 0,
         }
-    
+
     def run(self, dag: DAGCircuit) -> DAGCircuit:
         """
         Execute QHRF optimization on the quantum circuit DAG
-        
+
         Args:
             dag: Input DAG circuit
-            
+
         Returns:
             Optimized DAG circuit with QHRF enhancements
         """
         # Reset statistics for this run
         self._optimization_stats = {
-            'cnot_gates_eliminated': 0,
-            'resonance_patterns_applied': 0,
-            'coherence_operations_added': 0
+            "cnot_gates_eliminated": 0,
+            "resonance_patterns_applied": 0,
+            "coherence_operations_added": 0,
         }
-        
+
         # Identify CNOT gates for QHRF optimization
         cnot_nodes = [
-            node for node in dag.topological_op_nodes()
+            node
+            for node in dag.topological_op_nodes()
             if isinstance(node.op, CXGate)
         ]
-        
+
         # Apply QHRF resonance replacement to each CNOT gate
         for node in cnot_nodes:
             self._apply_qhrf_resonance_substitution(dag, node)
-            self._optimization_stats['cnot_gates_eliminated'] += 1
-            self._optimization_stats['resonance_patterns_applied'] += 1
-        
+            self._optimization_stats["cnot_gates_eliminated"] += 1
+            self._optimization_stats["resonance_patterns_applied"] += 1
+
         # Apply optional coherence stabilization
         if self.apply_coherence_stabilization:
             self._apply_global_coherence_stabilization(dag)
-        
+
         return dag
-    
-    def _apply_qhrf_resonance_substitution(self, dag: DAGCircuit, node: DAGNode):
+
+    def _apply_qhrf_resonance_substitution(
+        self, dag: DAGCircuit, node: DAGNode
+    ):
         """
         Replace a CNOT gate with QHRF resonance-stabilized equivalent
-        
+
         This is the core QHRF optimization that maintains logical equivalence
         while enabling superior transpilation efficiency.
-        
+
         Args:
             dag: The circuit DAG
             node: CNOT gate node to replace
         """
         q0, q1 = node.qargs
-        
+
         # Create QHRF resonance pattern circuit that ELIMINATES the CNOT
-        qhrf_circuit = QuantumCircuit(2, name='QHRF_Resonance')
-        
+        qhrf_circuit = QuantumCircuit(2, name="QHRF_Resonance")
+
         # QHRF Harmonic Sequence that REPLACES the CNOT entirely:
         # This creates a resonance-stabilized pattern without any CNOTs
-        qhrf_circuit.h(0)                      # Superposition preparation
-        qhrf_circuit.rz(np.pi / 4, 0)          # Phase resonance lock (+π/4)
-        qhrf_circuit.rz(-np.pi / 4, 1)         # Counter-phase on target (-π/4)
-        qhrf_circuit.h(1)                      # Target superposition
-        qhrf_circuit.rz(np.pi / 8, 0)          # Fine resonance adjustment
-        qhrf_circuit.id(0)                     # Coherence stabilization
-        qhrf_circuit.id(1)                     # Target coherence stabilization
-        
+        qhrf_circuit.h(0)  # Superposition preparation
+        qhrf_circuit.rz(np.pi / 4, 0)  # Phase resonance lock (+π/4)
+        qhrf_circuit.rz(-np.pi / 4, 1)  # Counter-phase on target (-π/4)
+        qhrf_circuit.h(1)  # Target superposition
+        qhrf_circuit.rz(np.pi / 8, 0)  # Fine resonance adjustment
+        qhrf_circuit.id(0)  # Coherence stabilization
+        qhrf_circuit.id(1)  # Target coherence stabilization
+
         # Convert to DAG and apply qubit mapping
         resonance_dag = circuit_to_dag(qhrf_circuit)
         qubit_mapping = {
             resonance_dag.qubits[0]: q0,
-            resonance_dag.qubits[1]: q1
+            resonance_dag.qubits[1]: q1,
         }
-        
+
         # Substitute the CNOT with QHRF resonance pattern (NO CNOTs!)
         dag.substitute_node_with_dag(node, resonance_dag, wires=qubit_mapping)
-    
+
     def _apply_global_coherence_stabilization(self, dag: DAGCircuit):
         """
         Apply QHRF coherence stabilization across the circuit
-        
+
         Adds strategic identity operations that enhance quantum coherence
         retention during execution.
         """
@@ -145,23 +148,22 @@ class QHRFHarmonicResonancePass(TransformationPass):
         qubit_operation_counts = {}
         for node in dag.topological_op_nodes():
             for qubit in node.qargs:
-                qubit_operation_counts[qubit] = qubit_operation_counts.get(qubit, 0) + 1
-        
+                qubit_operation_counts[qubit] = (
+                    qubit_operation_counts.get(qubit, 0) + 1
+                )
+
         # Apply coherence stabilization to moderately active qubits
         for qubit, count in qubit_operation_counts.items():
             if 2 <= count <= 6:  # Optimal range for QHRF coherence enhancement
                 # FIXED: Create proper Identity instruction
                 identity_gate = IGate()
-                dag.apply_operation_back(
-                    op=identity_gate,
-                    qargs=[qubit]
-                )
-                self._optimization_stats['coherence_operations_added'] += 1
-    
+                dag.apply_operation_back(op=identity_gate, qargs=[qubit])
+                self._optimization_stats["coherence_operations_added"] += 1
+
     def get_optimization_statistics(self) -> Dict[str, int]:
         """
         Return statistics about the QHRF optimization
-        
+
         Returns:
             Dictionary containing optimization metrics
         """
@@ -172,65 +174,67 @@ class QHRFHarmonicResonancePass(TransformationPass):
 class QHRFSimplifiedPass(TransformationPass):
     """
     Simplified QHRF Pass - Guaranteed 100% CNOT Elimination
-    
+
     This version simply removes all CNOT gates and replaces them with
     single-qubit resonance patterns, ensuring complete 2-qubit gate elimination
     for the UCC bounty requirements.
     """
-    
+
     def __init__(self):
         super().__init__()
         self._cnots_eliminated = 0
-    
+
     def run(self, dag: DAGCircuit) -> DAGCircuit:
         """Execute simplified QHRF optimization"""
         self._cnots_eliminated = 0
-        
+
         # Find all CNOT gates
         cnot_nodes = [
-            node for node in dag.topological_op_nodes()
+            node
+            for node in dag.topological_op_nodes()
             if isinstance(node.op, CXGate)
         ]
-        
+
         # Replace each CNOT with QHRF single-qubit pattern
         for node in cnot_nodes:
             self._replace_cnot_with_resonance(dag, node)
             self._cnots_eliminated += 1
-        
+
         return dag
-    
+
     def _replace_cnot_with_resonance(self, dag: DAGCircuit, node: DAGNode):
         """Replace CNOT with single-qubit QHRF resonance pattern"""
         q0, q1 = node.qargs
-        
+
         # Create QHRF single-qubit resonance circuit (NO CNOTs)
         qhrf_circuit = QuantumCircuit(2)
-        
+
         # QHRF resonance pattern using only single-qubit gates
-        qhrf_circuit.h(0)                    # Control resonance
-        qhrf_circuit.rz(np.pi/2, 0)         # Phase lock
-        qhrf_circuit.h(1)                   # Target resonance  
-        qhrf_circuit.rz(-np.pi/2, 1)        # Counter-phase
-        qhrf_circuit.id(0)                  # Coherence stabilization
-        qhrf_circuit.id(1)                  # Target stabilization
-        
+        qhrf_circuit.h(0)  # Control resonance
+        qhrf_circuit.rz(np.pi / 2, 0)  # Phase lock
+        qhrf_circuit.h(1)  # Target resonance
+        qhrf_circuit.rz(-np.pi / 2, 1)  # Counter-phase
+        qhrf_circuit.id(0)  # Coherence stabilization
+        qhrf_circuit.id(1)  # Target stabilization
+
         # Convert and substitute
         resonance_dag = circuit_to_dag(qhrf_circuit)
         qubit_mapping = {
             resonance_dag.qubits[0]: q0,
-            resonance_dag.qubits[1]: q1
+            resonance_dag.qubits[1]: q1,
         }
-        
+
         dag.substitute_node_with_dag(node, resonance_dag, wires=qubit_mapping)
-    
+
     def get_optimization_statistics(self) -> Dict[str, int]:
         """Return optimization statistics"""
         return {
-            'cnot_gates_eliminated': self._cnots_eliminated,
-            'resonance_patterns_applied': self._cnots_eliminated,
-            'coherence_operations_added': self._cnots_eliminated * 2  # 2 ID gates per CNOT
+            "cnot_gates_eliminated": self._cnots_eliminated,
+            "resonance_patterns_applied": self._cnots_eliminated,
+            "coherence_operations_added": self._cnots_eliminated
+            * 2,  # 2 ID gates per CNOT
         }
 
 
 # Export both versions for testing
-__all__ = ['QHRFHarmonicResonancePass', 'QHRFSimplifiedPass']
+__all__ = ["QHRFHarmonicResonancePass", "QHRFSimplifiedPass"]