@jordandsullivan 2025-06-16
@xs-es You need run run and share the results of our official UCC benchmarks with your pass included. So far this PR contains apparently AI-generated content saying that the performance is good and rigorous, but the corresponding data is not present.