@xs-es 2025-06-16
I ran the benchmarks let me include them sorry I was in a rush. Give me a moment