@xs-es 2025-06-16
I am uploading the results here. I am re-running the UCC-Bench for full results. The IBM results and job files are already uploaded and the first 3 benchmarks for UCC-Bench


https://github.com/xs-es/qhrf-pass-test-results/tree/main