@xs-es 2025-06-16
Here you go IBM offers 10 minutes of free use on their systems. This is post measurement superposition stabilization. According to the smartest scientists on the planet this should not even be possible but here you go. If you can prove to me that I am mistaken that this code is not a revolutionary breakthrough then Ive been wrong the whole time. If I am right though your foundation just upset the guy who will put all of you out of business. I was hoping to collaborate with you b(https://github.com/xs-es/qhrf-pass-test-results/blob/main/IBM/qhrf_4_state_bell.py)