@xs-es 2025-06-16
Im wondering how you reviewed 50 papers in 2 minutes.