@xs-es 2025-06-16
Im wondering how you reviewed 50 papers in 2 minutes. 

Heres 40 pages of "No Research" just one of my papers and yes they would have my name on them have you ever read a scientific study. Wow this is a real joke 

https://www.academia.edu/resource/work/129862344