@xs-es 2025-06-16
I'm right and I know I'm right but just imagine for a second if you thought I was right if you actually took the time to read the studies and realized oh wait a second he's on to something imagine how much money your foundation would make and you just passed it up I mean that alone is astonishing I even gave you the code to run on IBM quantum computer and I bet you won't even do it you won't even take the 5 minutes to run the code and that just cost you what billions of dollars and I was going to give it to you for 500 bucks wow glad I liked it