@jordandsullivan 2025-06-16
@xs-es I can see you're frustrated. Unfortunately though, your contribution here is not up to par with our standards. I would recommend again that if you believe you have achieved a genuine breakthrough, to submit your findings through a peer review process in a respected conference or journal. 

Please respect our Unitary Foundation Community Guidelines by taking in this feedback gracefully. 
https://ucc.readthedocs.io/en/latest/CODE_OF_CONDUCT.html