@xs-es 2025-06-16
I put alot of work into this and I feel like you are not even taking the time to properly review my work. My framework and everything else aside, have I not completed the requirements for the bounty by reducing gate count and circuit depth?