@xs-es 2025-06-18
I clearly explained this before hand I do not understand why I second guess myself. You approved it I  did it and now you are making up excuses. Not only did I destroy your bounty I revolutionized quantum physics in the process.  And here is the proof. After measurement you can clearly see the bell state maintain 4 states meaning I preveneted collapse of an entangled state even after measurement. This can then be recycled and used again significantly reducing overhead. 

[[NEW PASS] QHRF Phase Lock Optimization · unitaryfoundation_ucc · Discussion #406.pdf](https://github.com/user-attachments/files/20794970/NEW.PASS.QHRF.Phase.Lock.Optimization.unitaryfoundation_ucc.Discussion.406.pdf)
[qhrf_vs_CNOT.pdf](https://github.com/user-attachments/files/20795104/qhrf_vs_CNOT.pdf)